refactor: remove redundant code
