+++ packages/runtime/src/useCases/transport/relationships/CreateRelationship.ts
@@ -1,14 +1,13 @@
 import { Serializable } from "@js-soft/ts-serval";
-import { ApplicationError, Result } from "@js-soft/ts-utils";
-import { IncomingRequestsController, LocalRequestStatus } from "@nmshd/consumption";
-import { RelationshipCreationContent } from "@nmshd/content";
+import { Result } from "@js-soft/ts-utils";
+import { IncomingRequestsController } from "@nmshd/consumption";
 import { CoreId } from "@nmshd/core-types";
 import { AccountController, RelationshipTemplate, RelationshipTemplateController, RelationshipsController } from "@nmshd/transport";
 import { Inject } from "@nmshd/typescript-ioc";
 import { RelationshipDTO } from "../../../types";
 import { RelationshipTemplateIdString, RuntimeErrors, SchemaRepository, SchemaValidator, UseCase } from "../../common";
 import { RelationshipMapper } from "./RelationshipMapper";
-import { validateTypeOfCreationContentOfRelationship } from "./utility/validateTypeOfCreationContentOfRelationship";
+import { validateCreationContentOfRelationship } from "./utility/validateCreationContentOfRelationship";
 
 export interface CreateRelationshipRequest {
     templateId: RelationshipTemplateIdString;
@@ -42,41 +41,15 @@ export class CreateRelationshipUseCase extends UseCase<CreateRelationshipRequest
         const canSendRelationshipResult = await this.relationshipsController.canSendRelationship({ creationContent: request.creationContent, template });
         if (!canSendRelationshipResult.isSuccess) throw canSendRelationshipResult.error;
 
-        const typeOfCreationContentOfRelationshipValidationError = validateTypeOfCreationContentOfRelationship(template, request.creationContent);
-        if (typeOfCreationContentOfRelationshipValidationError) return Result.fail(typeOfCreationContentOfRelationshipValidationError);
+        const creationContentOfRelationshipValidationError = await validateCreationContentOfRelationship(this.incomingRequestsController, template, request.creationContent);
+        if (creationContentOfRelationshipValidationError) return Result.fail(creationContentOfRelationshipValidationError);
 
         const transformedCreationContent = Serializable.fromUnknown(request.creationContent);
-
-        if (transformedCreationContent instanceof RelationshipCreationContent) {
-            const responseToRequestOfTemplateValidationError = await this.validateResponseToRequestOfTemplate(request.templateId, transformedCreationContent);
-            if (responseToRequestOfTemplateValidationError) return Result.fail(responseToRequestOfTemplateValidationError);
-        }
 
         const sendRelationshipResult = await this.relationshipsController.sendRelationship({ template, creationContent: transformedCreationContent.toJSON() });
 
         await this.accountController.syncDatawallet();
 
         return Result.ok(RelationshipMapper.toRelationshipDTO(sendRelationshipResult));
-    }
-
-    private async validateResponseToRequestOfTemplate(
-        templateId: RelationshipTemplateIdString,
-        relationshipCreationContent: RelationshipCreationContent
-    ): Promise<ApplicationError | undefined> {
-        const acceptedIncomingRequests = await this.incomingRequestsController.getIncomingRequests({
-            status: LocalRequestStatus.Decided,
-            "source.reference": templateId,
-            "response.content.result": "Accepted"
-        });
-
-        if (acceptedIncomingRequests.length === 0) {
-            return RuntimeErrors.relationships.noAcceptedIncomingRequest();
-        }
-
-        if (acceptedIncomingRequests[0].response!.content.serialize() !== relationshipCreationContent.response.serialize()) {
-            return RuntimeErrors.relationships.wrongResponseProvidedAsCreationContent();
-        }
-
-        return;
     }
 }
